% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion.R
\name{fit_dispersion}
\alias{fit_dispersion}
\title{Fit Dispersion Parameter for Negative Binomial Model}
\usage{
fit_dispersion(
  beta,
  model_matrix,
  y,
  offset_matrix,
  tolerance,
  max_iter,
  do_cox_reid_adjustment = TRUE
)
}
\arguments{
\item{beta}{Vector of coefficient estimates}

\item{model_matrix}{Design matrix of predictor variables}

\item{y}{Vector of response variables (counts)}

\item{offset_matrix}{Matrix of offset values}

\item{tolerance}{Convergence tolerance for optimization}

\item{max_iter}{Maximum number of iterations for optimization}

\item{do_cox_reid_adjustment}{Logical indicating whether to apply Cox-Reid adjustment}
}
\value{
Estimated dispersion parameter
}
\description{
Estimates the dispersion parameter in a negative binomial GLM using
maximum likelihood estimation. Implementation from the glmGamPoi package.
}
\details{
This implementation comes from the glmGamPoi package:
https://github.com/const-ae/glmGamPoi

Original publication:
Ahlmann-Eltze, C., Huber, W. (2020). glmGamPoi: Fitting Gamma-Poisson
Generalized Linear Models on Single Cell Count Data. Bioinformatics.
https://doi.org/10.1093/bioinformatics/btaa1009
}
