% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{init_beta}
\alias{init_beta}
\title{Initialize Beta Coefficients Using Design Matrix}
\usage{
init_beta(y, design_matrix, offset_matrix)
}
\arguments{
\item{y}{Count data matrix}

\item{design_matrix}{Matrix of predictor variables}

\item{offset_matrix}{Matrix of offset values}
}
\value{
Matrix of initial beta coefficients
}
\description{
Initializes regression coefficients using QR decomposition
of the design matrix and normalized log counts.
}
