% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{init_beta_groups}
\alias{init_beta_groups}
\title{Initialize Beta Coefficients Using Groups}
\usage{
init_beta_groups(y, groups, offset_matrix)
}
\arguments{
\item{y}{Count data matrix}

\item{groups}{Vector indicating group membership}

\item{offset_matrix}{Matrix of offset values}
}
\value{
Matrix of initial beta coefficients by group
}
\description{
Initializes regression coefficients based on group-wise
means of normalized counts.
}
\keyword{internal}
