% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devil-package.R
\docType{package}
\name{devil-package}
\alias{devil}
\alias{devil-package}
\title{devil: devil}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

devil is an R package for fast and scalable single-cell differential expression analysis using sandwich covariance estimators to account for multi-patient variability, with optional GPU acceleration for efficient inference.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://caravagnalab.github.io/devil/}
  \item Report bugs at \url{https://github.com/caravagnalab/devil/issues}
}

}
\author{
\strong{Maintainer}: Giovanni Santacatterina \email{santacatterinagiovanni@gmail.com}

}
\keyword{internal}
