% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devil-package.R
\docType{package}
\name{devil-package}
\alias{devil}
\alias{devil-package}
\title{devil: devil}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

devil is an R package for fast, scalable differential expression analysis of single-cell and single-nucleus RNA-sequencing data. It is designed to operate directly at the cell level while accounting for complex experimental structure. It employs a statistical framework based on generalized linear models with cluster-robust (sandwich) covariance estimators, enabling valid inference in the presence of multi-patient, multi-sample, or repeated-measurement designs. To support analyses at atlas scale, devil combines efficient variational inference with optional GPU acceleration, delivering substantial gains in runtime and memory efficiency while retaining accurate and well-calibrated statistical tests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://caravagnalab.github.io/devil/}
  \item Report bugs at \url{https://github.com/caravagnalab/devil/issues}
}

}
\author{
\strong{Maintainer}: Giovanni Santacatterina \email{santacatterinagiovanni@gmail.com}

}
\keyword{internal}
