% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devil-package.R
\docType{package}
\name{devil-package}
\alias{devil}
\alias{devil-package}
\title{devil: devil}
\description{
An R package to perform fast and scalable single cell differential expression analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://caravagnalab.github.io/devil/}
}

}
\author{
\strong{Maintainer}: Giovanni Santacatterina \email{santacatterinagiovanni@gmail.com}

}
\keyword{internal}
