% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit_devil}
\alias{fit_devil}
\title{Fit Model Parameters}
\usage{
fit_devil(
  input_matrix,
  design_matrix,
  overdispersion = TRUE,
  offset = 0,
  size_factors = TRUE,
  verbose = FALSE,
  max_iter = 500,
  eps = 1e-04,
  parallel = TRUE
)
}
\arguments{
\item{input_matrix}{Vector of response variable.}

\item{design_matrix}{Matrix of predictor variables.}

\item{overdispersion}{Logical indicating whether to estimate overdispersion.}

\item{offset}{Offset vector to be included in the model. Defaults to 0.}

\item{size_factors}{Logical indicating whether to compute size factors. Defaults to TRUE.}

\item{verbose}{Logical indicating whether to display progress messages. Defaults to FALSE.}

\item{max_iter}{Maximum number of iterations for optimization. Defaults to 500.}

\item{eps}{Tolerance level for convergence criterion. Defaults to 1e-4.}

\item{parallel}{Logical indicating whether to use parallel computation. Defaults to TRUE.}
}
\value{
List containing fitted model parameters including beta coefficients,
overdispersion parameter (if estimated), and beta sigma.
}
\description{
This function fits model parameters including beta coefficients, dispersion parameter,
and beta sigma using provided predictor variables (input_matrix) and response variable (design_matrix).
It optionally estimates overdispersion based on the fitted model.
}
\details{
This function fits model parameters including beta coefficients, dispersion parameter,
and beta sigma using provided predictor variables (X) and response variable (y).
It optionally estimates overdispersion based on the fitted model.
}
