% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit_devil}
\alias{fit_devil}
\title{Fit Model Parameters}
\usage{
fit_devil(
  input_matrix,
  design_matrix,
  overdispersion = TRUE,
  offset = 0,
  size_factors = TRUE,
  verbose = FALSE,
  max_iter = 500,
  tolerance = 0.001,
  eps = 1e-06,
  CUDA = FALSE,
  batch_size = 1024L,
  parallel.cores = NULL
)
}
\arguments{
\item{input_matrix}{A numeric matrix representing the response variable, with rows corresponding to genes and columns to samples.}

\item{design_matrix}{A numeric matrix representing the predictor variables, with rows corresponding to samples and columns to predictors.}

\item{overdispersion}{Logical value indicating whether to estimate the overdispersion parameter. (default is \code{TRUE})}

\item{offset}{A numeric vector to be included as an offset in the model. (default is \code{0})}

\item{size_factors}{Logical value indicating whether to compute size factors for normalization. (default is \code{TRUE})}

\item{verbose}{Logical value indicating whether to display progress messages during execution. (default is \code{FALSE})}

\item{max_iter}{Integer specifying the maximum number of iterations allowed for the optimization process. (default is \code{500})}

\item{tolerance}{Numeric value indicating the tolerance level for the convergence criterion. (default is \code{1e-3})}

\item{eps}{A small numeric value added to \code{input_matrix} to avoid issues with non-invertible matrices. (default is \code{1e-6})}

\item{CUDA}{Logical value indicating whether to use GPU version of the code (default is \code{FALSE})}

\item{batch_size}{Integer specifying the number of genes that will be fit in each batch if \code{CUDA = TRUE}. (default is 1024)}

\item{parallel.cores}{Integer specifying the number of CPU cores to use for parallelization. If \code{NULL}, the maximum number of available cores are used. (defaults is \code{NULL})}
}
\value{
A list containing the following elements:
\item{beta}{A matrix of fitted beta coefficients for each gene.}
\item{overdispersion}{A numeric vector of overdispersion parameters for each gene (if estimated).}
\item{iterations}{A numeric vector indicating the number of iterations taken for each gene.}
\item{size_factors}{A numeric vector of size factors used for normalization.}
\item{offset_matrix}{A numeric matrix of offset values used in the model.}
\item{design_matrix}{The design matrix provided as input.}
\item{input_matrix}{The input matrix used after processing.}
\item{input_parameters}{A list of input parameters used in the function, including \code{max_iter}, \code{tolerance}, and \code{parallel.cores}.}
}
\description{
\code{fit_devil} performs model fitting by estimating beta coefficients, dispersion parameters,
and beta sigma. The function uses predictor variables provided in the \code{design_matrix} and a response
variable provided in the \code{input_matrix}. Optional features include the estimation of overdispersion
and the computation of size factors. The function supports parallel processing and allows customization
of various parameters such as the number of iterations, convergence tolerance, and more.
}
\details{
This function fits model parameters, including beta coefficients, the dispersion parameter,
and beta sigma, using the provided predictor variables (\code{design_matrix}) and response variable (\code{input_matrix}).
It optionally estimates overdispersion based on the fitted model.
}
