% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot}
\usage{
plot_volcano(
  devil.res,
  lfc_cut = 1,
  pval_cut = 0.05,
  labels = TRUE,
  colors = c("gray", "forestgreen", "steelblue", "indianred"),
  color_alpha = 0.7,
  point_size = 1,
  center = TRUE,
  title = "Volcano plot"
)
}
\arguments{
\item{devil.res}{A data frame or tibble containing the results of differential expression analysis. This object is generated by the \code{test_de} function.}

\item{lfc_cut}{A numeric value specifying the threshold for absolute log-fold change. Genes with absolute log-fold change greater than or equal to this value are highlighted. (default is \code{1})}

\item{pval_cut}{A numeric value specifying the threshold for adjusted p-value. Genes with adjusted p-value less than or equal to this value are highlighted. (default is \code{0.05})}

\item{labels}{Logical indicating whether to label significant genes on the plot. (default is \code{TRUE})}

\item{colors}{A character vector specifying the colors to use for different classes of genes (e.g., "non-significant", "lfc", "p-value", "p-value and lfc").
Default colors are \code{"gray"}, \code{"forestgreen"}, \code{"steelblue"}, and \code{"indianred"}}

\item{color_alpha}{A numeric value specifying the alpha transparency level for point colors. (default is \code{0.7})}

\item{point_size}{A numeric value specifying the size of points in the plot. (default is \code{1})}

\item{center}{Logical indicating whether to center the x-axis at zero. (default is \code{TRUE})}

\item{title}{A character string specifying the title of the plot. (default is \verb{Volcano plot})}
}
\value{
A \code{ggplot} object representing the volcano plot.
}
\description{
Plot a volcano plot for differential expression analysis.
}
\details{
This function creates a volcano plot for visualizing the results of differential expression analysis. It highlights genes based on their log-fold change (\code{lfc}) and adjusted p-values (\code{adj_pval}). Genes that meet the specified thresholds for both log-fold change and adjusted p-value are optionally labeled on the plot.

The function also ensures that genes with an adjusted p-value of zero are displayed with a minimal p-value (\code{1e-16}) to avoid plotting issues.
}
