% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot}
\usage{
plot_volcano(
  devil.res,
  lfc_cut = 1,
  pval_cut = 0.05,
  labels = TRUE,
  colors = c("gray", "forestgreen", "steelblue", "indianred"),
  color_alpha = 0.7,
  point_size = 1,
  center = TRUE,
  title = "Volcano plot"
)
}
\arguments{
\item{devil.res}{A data frame or tibble containing the results of differential expression analysis of the function \code{test_de}.}

\item{lfc_cut}{The threshold for absolute log-fold change. Genes with absolute log-fold change greater than or equal to this value are highlighted.}

\item{pval_cut}{The threshold for adjusted p-value. Genes with adjusted p-value less than or equal to this value are highlighted.}

\item{labels}{Logical indicating whether to label significant genes on the plot.}

\item{colors}{A vector of colors to use for different classes of genes.}

\item{color_alpha}{The alpha value for point colors.}

\item{point_size}{The size of points in the plot.}

\item{center}{Logical indicating whether to center the x-axis at zero.}

\item{title}{The title of the plot.}
}
\value{
A ggplot object representing the volcano plot.
}
\description{
Plot a volcano plot for differential expression analysis.
}
\details{
This function creates a volcano plot for visualizing differential expression analysis results.
It highlights genes based on their log-fold change and adjusted p-values.
Genes meeting the specified thresholds for both log-fold change and adjusted p-value are labeled.
}
